+++ aether/e2e-tests/src/test/java/org/pragmatica/aether/e2e/ControllerE2ETest.java
@@ -46,7 +46,7 @@ class ControllerE2ETest {
 
     @BeforeAll
     static void createCluster() {
-        cluster = AetherCluster.aetherCluster(3, PROJECT_ROOT);
+        cluster = AetherCluster.aetherCluster(5, PROJECT_ROOT);
         cluster.start();
         cluster.awaitQuorum();
         cluster.awaitAllHealthy();