+++ aether/e2e-tests/src/test/java/org/pragmatica/aether/e2e/NetworkPartitionE2ETest.java
@@ -48,7 +48,7 @@ class NetworkPartitionE2ETest {
 
     @BeforeAll
     static void createCluster() {
-        cluster = AetherCluster.aetherCluster(3, PROJECT_ROOT);
+        cluster = AetherCluster.aetherCluster(5, PROJECT_ROOT);
         cluster.start();
         cluster.awaitQuorum();
         cluster.awaitAllHealthy();